% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_file_availability}
\alias{check_file_availability}
\title{Utility funciton to check if File is available from URL}
\usage{
check_file_availability(file_url)
}
\arguments{
\item{file_url}{string pointing to OGD Ressource URL}
}
\value{
boolean if file is available
}
\description{
Utility funciton to check if File is available from URL
}
