% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_full_load}
\alias{check_full_load}
\title{function to check if files in folder and ogd_messorte are same (full load)}
\usage{
check_full_load(csv_paths_list_all_einzel, csv_paths_list_all_langzeit)
}
\arguments{
\item{csv_paths_list_all_einzel}{list with paths to csv files for Einzelmessungen}

\item{csv_paths_list_all_langzeit}{list with paths to csv files for Langzeitmessungen}
}
\value{
console output if everything went fine. Fails informatively if differences occur.
}
\description{
function to check if files in folder and ogd_messorte are same (full load)
}
