% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_na_einzelmessungen}
\alias{check_na_einzelmessungen}
\title{Utility function to check for NAs (Einzelmessungen)}
\usage{
check_na_einzelmessungen(einzelmessungen_list)
}
\arguments{
\item{einzelmessungen_list}{list with paths to csv files}
}
\value{
if any NAs are found: list with IDs
}
\description{
Utility function to check for NAs (Einzelmessungen)
}
