% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_na_langzeitmessungen}
\alias{check_na_langzeitmessungen}
\title{Utility function to check for NAs (Langzeitmessungen)}
\usage{
check_na_langzeitmessungen(langzeitmessungen_list)
}
\arguments{
\item{langzeitmessungen_list}{list with paths to csv files}
}
\value{
if any NAs are found: list with IDs
}
\description{
Utility function to check for NAs (Langzeitmessungen)
}
