% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_data.R
\name{clean_langzeit_list}
\alias{clean_langzeit_list}
\title{function to delete already available measurement IDs from a list (Langzeitmessungen)}
\usage{
clean_langzeit_list(csv_list, df_nested_list)
}
\arguments{
\item{csv_list}{list with folder/path names}

\item{df_nested_list}{list with ids (from OGD ressource)}
}
\value{
a "cleaned" list which only contains IDs which are not already present in the OGD ressource
}
\description{
function to delete already available measurement IDs from a list (Langzeitmessungen)
}
