% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function_extraction.R
\name{extract}
\alias{extract}
\title{function to start the whole extraction process}
\usage{
extract(
  delta_load = TRUE,
  full_load = FALSE,
  path_rohdaten_topfolder,
  url_ogd_messwerte =
    "https://www.web.statistik.zh.ch/ogd/daten/ressourcen/KTZH_00002462_00004983.csv",
  url_ogd_messorte =
    "https://www.web.statistik.zh.ch/ogd/daten/ressourcen/KTZH_00002462_00004924.csv"
)
}
\arguments{
\item{delta_load}{boolean indicating if a delta load should be performed}

\item{full_load}{boolean indicating if a full load should be performed}

\item{path_rohdaten_topfolder}{character vector (string) pointing to folder where Measurement data is stored on local machine}

\item{url_ogd_messwerte}{character vector (string) representing URL to OGD Messwerte Ressource}

\item{url_ogd_messorte}{character vector (string) representing URL to OGD Messorte Ressource}
}
\value{
writes rohdaten_messwerte.csv into inst/extdata/temp/extract/
}
\description{
function to start the whole extraction process
}
