% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_csv_paths_einzel}
\alias{extract_csv_paths_einzel}
\title{function to generate a list of csv paths (Einzelmessungen)}
\usage{
extract_csv_paths_einzel(top_folder_path, num_dirs = Inf)
}
\arguments{
\item{top_folder_path}{character string representing the file path to the folder where Langzeitmessungen & Einzelmessunge are located}

\item{num_dirs}{number of (sub)-directories to process. Default = \code{Inf} which means all directories. Useful for testing purposes}
}
\value{
list with paths to csv files
}
\description{
function to generate a list of csv paths (Einzelmessungen)
}
