% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{filter_csv_files_einzel}
\alias{filter_csv_files_einzel}
\title{Function to filter out HEADER CSV files (Einzelmessungen)}
\usage{
filter_csv_files_einzel(paths_list)
}
\arguments{
\item{paths_list}{list with paths to csv files}
}
\value{
cleaned list with paths to csv files
}
\description{
Function to filter out HEADER CSV files (Einzelmessungen)
}
