% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_csv_functions.R
\name{process_csv_data_einzel}
\alias{process_csv_data_einzel}
\title{function to read in a list of csv paths (Einzelmessungen)}
\usage{
process_csv_data_einzel(filtered_paths_list)
}
\arguments{
\item{filtered_paths_list}{list with file paths to csv files - produced by \code{\link[=extract_csv_paths_einzel]{extract_csv_paths_einzel()}})}
}
\value{
nested list object which contains the actual measurement data as separate dataframes
}
\description{
function to read in a list of csv paths (Einzelmessungen)
}
