% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_data.R
\name{report_delta_langzeit}
\alias{report_delta_langzeit}
\title{function to report the delta between the existing ODG data and the measurement data in a folder (Langzeitmessungen)}
\usage{
report_delta_langzeit(existing_ogd, delta_langzeit)
}
\arguments{
\item{existing_ogd}{list with existing ogd - produced by \code{\link[=retrieve_existing_data]{retrieve_existing_data()}})}

\item{delta_langzeit}{list with delta langzeit (new Langzeit data in folder)}
}
\value{
invisbly returns a tibble with the delta - prints always output to the console
}
\description{
function to report the delta between the existing ODG data and the measurement data in a folder (Langzeitmessungen)
}
