% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_data.R
\name{retrieve_existing_data}
\alias{retrieve_existing_data}
\title{function to retrieve the existing OGD Data}
\usage{
retrieve_existing_data(
  messwerte_url =
    "https://www.web.statistik.zh.ch/ogd/daten/ressourcen/KTZH_00002462_00004983.csv",
  messorte_url =
    "https://www.web.statistik.zh.ch/ogd/daten/ressourcen/KTZH_00002462_00004924.csv"
)
}
\arguments{
\item{messwerte_url}{character vector with URL pointing to OGD ressource "Aufbereitete Messwerte"}

\item{messorte_url}{character vector with URL pointing to OGD ressource "Messorte"}
}
\value{
list with three elements: "Einzelmessungen", "Langzeitmessungen" & Messorte. All objects are dataframes and are needed for the delta check
}
\description{
function to retrieve the existing OGD Data
}
