% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update MDV Messwerte Ressourcen der Immissionsmessungen NIS im Kanton Zürich als OGD
Dieses Skript eignet sich um die OGD Schulhausmessungsdaten (nur Messdaten) zu aktualisieren.
Achtung: Zur Aktualiserung des 'Messorte' CSV File -> über das GUI des MDV hochladen. Mit diesem Skript werden nur
die Messwerte Files (Rohdaten Messwerte & Aufbereitete Messwerte) aktualisiert.
Das OGD Dataset muss bereits in der Metadatenverwaltung angelegt sein und die richtige ID in den Parameter (siehe weiter unten) angegeben werden.
Es braucht Schreibrechte im MDV, welche mit den dazugehörigen Username Passwort Kombination möglich ist (siehe nächste Linie). Ausserdem braucht man einen Token, um das neuste zhMetadatenAPI Package von Github herunterzuladen.
Bitte beachte, dass im .Renviron File die Variabeln: 'mdv_user', 'mdv_pw' und 'ZH_METADATEN_API_TOKEN' mit dazugehörigen Werten vorhanden ist.}
\usage{
update()
}
\description{
Update MDV Messwerte Ressourcen der Immissionsmessungen NIS im Kanton Zürich als OGD
Dieses Skript eignet sich um die OGD Schulhausmessungsdaten (nur Messdaten) zu aktualisieren.
Achtung: Zur Aktualiserung des 'Messorte' CSV File -> über das GUI des MDV hochladen. Mit diesem Skript werden nur
die Messwerte Files (Rohdaten Messwerte & Aufbereitete Messwerte) aktualisiert.
Das OGD Dataset muss bereits in der Metadatenverwaltung angelegt sein und die richtige ID in den Parameter (siehe weiter unten) angegeben werden.
Es braucht Schreibrechte im MDV, welche mit den dazugehörigen Username Passwort Kombination möglich ist (siehe nächste Linie). Ausserdem braucht man einen Token, um das neuste zhMetadatenAPI Package von Github herunterzuladen.
Bitte beachte, dass im .Renviron File die Variabeln: 'mdv_user', 'mdv_pw' und 'ZH_METADATEN_API_TOKEN' mit dazugehörigen Werten vorhanden ist.
}
\examples{
update()

}
