% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Erstelle die beiden Messwerte OGD Files (rohdaten_messwerte.zip und aufbereitete_messwerte.csv)
f端r die Visualisierung der Immissionsmessungen NIS im Kanton Z端rich als OGD}
\usage{
transform(full_load = TRUE)
}
\description{
Erstelle die beiden Messwerte OGD Files (rohdaten_messwerte.zip und aufbereitete_messwerte.csv)
f端r die Visualisierung der Immissionsmessungen NIS im Kanton Z端rich als OGD
}
